Build dkml-installer-ocaml-offline
+build: dkml-component-ocamlcompiler-offline

+bugfix: Use -y (say yes) while running create-opam-switch

+bugfix: Install diskuv-opam-repository packages at
the same time as FULL_NOT_DUNE packages so
indirect dependencies of diskuv-opam-repository
packages like [base] are installed with the correct
version. Previously the latest base [base.v0.16.2]
was broken in MSVC, and since [feather] was in
diskuv-opam-repository and [feather] had a dependency
on [base], that caused the broken [base.v0.16.2]
to be installed. However, the good [base.v0.16.1]
was listed in dkml-runtime-distribution (ie.
FULL_NOT_DUNE packages). // As a bonus, we get
less re-installs because the package versions don't
switch from latest (ex. base.v0.16.2) back down to
the FULL_NOT_DUNE version (ex. base.v0.16.1)