bugfix: Trigger with SHORTREF not GLOBs
Change detection of _deps/*-src projects for Stage*
targets used to be based on file(GLOB) and
file(GLOB_RECURSE) expressions.
However that was:
- expensive (file scanning ... especially if _build or _opam
development was done on the _deps project)
- partial (since may miss a new file)

However if the target actually ran, a different check was
done on the git commit of the _deps project to see if an
[opam install] was executed.

This change now consolidates the change logic to just
the git commit of the _deps project.